memory_mcp/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
